Update references after moving CustomReplyEditorActivity
