+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
         </activity>
 
         <activity
-            android:name=".CustomReplyEditorActivity"
+            android:name=".activity.customreplyeditor.CustomReplyEditorActivity"
             android:label="@string/mainAutoReplyLabel"
             android:parentActivityName="com.parishod.wareply.activity.main.MainActivity" >
         </activity>
