+++ app/src/main/res/layout/activity_custom_reply_editor.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CustomReplyEditorActivity">
+    tools:context=".activity.customreplyeditor.CustomReplyEditorActivity">
 
     <com.google.android.material.textfield.TextInputLayout
         android:id="@+id/autoReplyTextInputLayout"
